CO-1114. Handle different worker-user-data secret name for 4.6.
The worker-user-data secret name became worker-user-data-managed in 4.6.
Hive now supports alternate names for this secret.

When the OpenShift version reported in the ClusterDeployment cannot
be parsed as a semver, then assume that the cluster is using the
new name for the worker user data secret. We know that legacy
OpenShift versions use semver, so we can conclude that any version
that is not using semver must not be a legacy version.