Implement HibernateAfter to auto-powerdown.
Added to ClusterDeployment.Spec.

When a cluster has been in a running state (installed with no hibernate
condition, or hibernate condition=false) for more than the
hibernate-after duration, the cluster will be automatically put to
sleep.

Intended for dev clusters where you do not want them to accidentally be
left running beyond the workday.