Fix metrics crash on new InstallLaunchError condition.
In 8b7044b12ee5de1cc9c54c46a91b9b946aefa854 a new ClusterDeployment
condition type was added, but it was not added to the list of
AllClusterDeploymentConditions.

When encountered in the wild the metrics code would attempt to increment
an int in a map we expected to be initialized to 0 but was instead nil.

This change adds the condition type to the list of all, adds a warning
comment, and a safety check and log warning if this happens again in the
future.