Fix failure to detect unreachable in ssi controller.
This particular controller uses a dynamic client which does not
automatically perform discovery API calls. In some or possibly all
scenarios, this controller was not able to detect and set the
unreachable condition. (which we normally assume the first controller to
detect it will set the condition)

This change adds a get for the kube-system namespace, which should
exist, but even if it doesn't we just check for any error that isn't
"not found". This should allow better unreachable detection if the ssi
controller is the first to see it.