controller: create clustersync as separate reconcile
For a new ClusterDeployment, the ClusterSync is created and then immediately
updated with the status of applying any syncsets. If there were errors applying
syncsets, then the clustersync controller may do its next reconcile using the
created version instead of the updated version. When this happens, the update
at the end of the reconcile will fail due to a conflict.

With these changes, when there is no ClusterSync, the clustersync controller will
create the ClusterSync as the only thing that it does for the reconciliation.
The clustersync controller will requeue the ClusterDeployment when the controller
sees the new ClusterSync, and the controller will not get any conflicts when doing
updates.