controller: requeue cd when csync is created
With commit 0c17526, the ClusterSync is
created as a separate clustersync reconcile. However, the clustersync
controller does not watch for changes to ClusterSync. Consequently, the
syncsets are not applied to a new cluster until there is some later change
to the ClusterDeployment or one of the syncsets. These changes will
queue the ClusterDeployment when the ClusterSync is created.