deprovision: kill deprovision when creds change
The "--creds-dir" flag has been added to the `hiveutil deprovision`
command. When this flag is used, the command will create a file
observer watching for any changes to any files in that directory.
If a change is detected, the command will be terminated.

The uninstall pods for platforms other than AWS have been updated
to include the "--creds-dir" flag in the call to the deprovision
command.

https://issues.redhat.com/browse/CO-1135