machinepool condition for unsupported spot instances
Add the UnsupportedConfiguration condition to the MachinePool type.
This condition is used to indicate that the cluster does not support
the configuration specified in the MachinePool. If the MachinePool
specifies spot instances but the cluster is running an OpenShift
version earlier than 4.5, then the UnsupportedConfiguration
condition will be set on the MachinePool and not MachineSets will
be generated.

https://issues.redhat.com/browse/CO-1042