send a non-nil list of Zones for OpenStack
The installer machineset generator will distribute all the replicas across the list of zones. If we do not provide a list of zones, we get back an empty list of machinesets. Follow what the installer does, and provide list of zones with exactly one member consisting of the empty string.