makefile: use earlier version of controller-gen
Force build-machinery-go to grab an earlier version of controller-gen.
The newer version adds annotations on pod specs that will cause the CRDs
to fail validation. The fix for that requires the CRDs to use v1, but we
still need to support users that are running kube versions that only have
v1beta1.

https://issues.redhat.com/browse/CO-1193