Fix stalled leader election after Deployment updates
With the current Deployment strategy of rolling update, a new pod is brought up
before the old is terminated. The new pod leader election has already determined
someone else is leader before the old releases it's lock, and will wait the full
renew interval of 270 seconds before checking again.

To fix we switch to the recreate strategy which will delete the old pod first,
then create the new. This seems the desired strategy for leader election based
operators.