Stop reporting SelectorSyncSet metrics after initial sync success.
Some SelectorSyncSet time to first apply metrics were abnormally large
as labels can be added to clusters long after they're installed,
resulting in a time to apply metric of install time -> now, which could
be days weeks or months.

Instead use the new "FirstSuccessTimestamp" on ClusterSync.Status. If we
see a first success for a SelectorSyncSet, but the FirstSuccessTimestamp
overall is set, we know this was a new match and thus we don't really
have an accurate start time, and the metric should be skipped.