Use time.Sleep for simulating apply scale tests.
Initial attempt at this used a network request to a small node.js server
that would be told how long to wait until returning. However if we were
to lookup real world request times to seed the random values for this,
we would then add the network overhead on top of our real world
requests, leading to a double simulation of the wait, which could be
somewhat less accurate.

The sleep allows us to use the values we can lookup from our metrics
hopefully more accurately.

In my testing the performance of the two seems fairly similar, and
appears to be as effective at simulating as using a real network call,
but without the overhead of standing up a nodejs service and thus
something we can more easily commit and make use of as developers.