Add dev prometheus config for scraping multiple pods.
This is in preparation for the hive-clustersync pods that we expect to
soon be running, we no longer will have one pod to scrape by going to a
service, rather we'll want to scrape all pods by endpoints behind the
service.

Breaks out hive-controllers (name matches app-sre configuration,
although it doesn't matter as this is just dev prometheus setup for
scale testing), and adds hive-clustersync job. Both are capable of
scraping multiple pods, though this is not strictly relevant for
hive-controllers, it should work fine.