apis/hive/v1: allow provinding a secret containing trust bundle for openstack APIs
To support self-signed OpenStack endpoints, which is very common, the hive configuration needs to
allow users to specify a trust bundle.

The trust needs to be updated in two locations,
1. The pod that runs the installer binary, and
2. OpenStack specific configuration that would make the components in OpenShift cluster to
   trust the specified certificate bundle.
These two steps are detailed here [1]

So following the oVirt [2] and vSphere [3] APIs, this PR adds `optional` certificatesSecretRef field to the cluster deployement
for OpenStack. When set this will update the root trust with certificate bundles provided in the referenced secret making sure Step 1
from above is achieved. But since Step 2 requires clouds.yaml provided by the user to reference a file from disk, this change makes the
location in the installer pod part of the API for cluster deployment. So the trust bundle will be available at
either `/etc/openstack-ca/<key name containing the trust bundle in credentialsSecretRef Secret>` or the
`/etc/pki/ca-trust/extracted/pem/tls-ca-bundle.pem` file path.

The certificatesSecretRef can be nil but when set it must always point to a non-empty secret name.

The host trust bundle update for installation / cleanup is same as vSphere/Ovirt [4], [5] while making sure this update is only invoked
when the certificate bundle is set.

[1]: https://github.com/openshift/installer/blob/master/docs/user/openstack/README.md#self-signed-openstack-ca-certificates
[2]: https://github.com/openshift/hive/blob/master/pkg/apis/hive/v1/ovirt/platform.go#L14-L16
[3]: https://github.com/openshift/hive/blob/master/pkg/apis/hive/v1/vsphere/platform.go#L16-L18
[4]: https://github.com/openshift/hive/blob/master/pkg/install/generate.go#L338-L341
[5]: https://github.com/openshift/hive/blob/master/pkg/install/generate.go#L721-L732