Fix control plane cert patching for new clusters.
Recent change was not tested with new clusters as all eyes were on
making sure we rolled out for existing clusters smoothly.

JSON patches will not apply if the json sections referenced do not yet
exist.

This change adds two new directions to the json patch to ensure that the
two parent sections for what we need to patch exist. Because these are
all within the same json patch, they do not modify the resource. (which
they would if we applied each individually)