Report DNSNotReady in the cluster provisioning underway metric.
Update the metric to only use conditions if their status is true
indicating an unexpected state.

Add DNSNotReady at the start of the list, so that if this is true it is
the reason we report without checking the others.

Similarly reorder provision launch error as this is likely to surface
before an install failure.