Fix resuming clusters stuck in FailedToStart.
In the event clusters fail to start (i.e. AWS instance capacity issues
we've seen the last few days), Hive will mark the Hibernation condition
reason FailedToStart, but then would stop all processing, there was no
way to get out of this state, or resume the cluster again.

This change adds the FailedToStart reason to the case where we continue
trying to start machines. If an error is encountered starting the
machines, we also now return this error from the controller, which gives
us the normal reconcile with backoff loop we want in a situation like
this.

In practice with the AWS capacity issues this week, this situation will
actually resolve itself fairly quickly if you keep trying.

Also allow users to transition clusters back to Hibernation power state if
stuck in FailedToStart.