Add annotation on ClusterDeployment that will fake install/apply.
To better be able to simulate the etcd load of clusters, this change
switches from adopting (which bypasses significant etcd storage for the
ClusterProvision, Job, and Pod all of which carry a podspec), and
instead will run a quick install pod that skips the openshift-install
execution and just returns the admin kubeconfig specified in the
ClusterDeployment annotation.

The hive.openshift.io/fake-install-kubeconfig-secret annotation should
be a Secret name which exists in the same namespace as the
ClusterDeployment. The caller must provide this, in this case our
test_setup.sh script.