clusterpool: include deleting claimed clusters for max concurrent
The cluster pool 's max concurrent will include the deleting claimed clsuters. This is
to make sure cluster pool actions are safe and under limits as specified by the admin.

previously the claimed clusters from a pool, were deleted out band directly by the clusterclaim
controller when the claim was deleted. This made controlling the max concurrent provision/
deprovision difficult because 2 controllers are making the same decision making independently.
To enforce the max concurrent for pool, all the actions of cluster provision/deprovision had to
be moved to clusterpool controller so that it can ensure the max concurrent limit.

Thefore, the clusterclaim controller now annotates the clusterdeployment that it should be removed
when the claim is deleted, it still removes all the other claim specific resources from the the
namespace like rolebindings etc, and removes the finalizer. The clusterpool controller removes
this previously claimed cluster whenever there is budget to do so.

Since clusters in a pool that have been claimed can live after the cluster pool has been deleted,
we need to make sure these clusters are also cleaned up when the claims are deleted. To fulfil that
the clusterpoolnamespace controller will cleanup up all clusters that belonged to a clusterpool that
no longer exists and have been annotated by the claim as safe to remove.