Allow public subnets in AWS MachinePool
To be consistent with install config, allow public subnets to be
specified in MachinePool even if the MachineSets created would use only
private subnets.

Account for success in all of the following cases:
- No subnets specified
- Only private subnets specified
- Both public and private subnets specified

Private subnets are required to be exactly one per availability zone if
subnets are specified. If public subnets are also specified, then there
should be exactly 1 public and 1 private subnet for each availability
zone.