use ClusterDeployment's CA data when validating vsphere credentials
When a vSphere ClusterDeployment defines the CA certificates Secret, use the CA certificate data when validating vSphere credentials.

The high-level client library doesn't allow providing a CA certificate early enough, so we need to implement a lot of the lower-level connection setup so that we can slip in the custom CA data when needed.