clusterpool: allow setting lifetime for claims
claimLifetime allows admins to set the lifetime for all cluster claims created for
a cluster pool. Two new lifetime configurations are available, `default` and `maximum`

clusterpool`.spec.cliamLifetime.default`
This sets the default lifetime for the claim when one is not set by the claim itself.

clusterpool`.spec.claimLifetime.maximum`
This sets the maximum lifetime for all claims in the cluster pool. If any user sets lifetime
on the claim, the actual lifetime will be minimum of the maximum allowed by the pool and what was
set on the claim.
This allows the admins makes sure that the tightest lifetime constraint is always respected for the
cluster pool.

Also a new status field is added to the clusterclaim `.status.lifetime` that provides information of
the exact lifetime including all the pool configurations and the claim spec.

NOTE: the cluster pool controller could also set the lifetime on the claim when it assigns a cluster
to the claim but that would would require setting the intention on the object as part of the UPDATE hence
making it difficult to
- later figure out if the lifetime was set by user or pool which is important if the claim ever had to
move to a different pool.
- also the user could change the claim lifetime to somthing else after the controller updated it and the controller
would have to fiure out if the change should be kept and then revert the change, possibly race with the user.
This current approach with clusterclaim controller computing lifetime allows users to express their intentions at
both levels in a much more clear way and it can be enforced similarly.