remotemachineset: update the trust from CertificatesSecretRef for generating machinesets
Previously the controller would use the installer to generate the machinesets. That installer
function communicates with OpenStack APIs using the clouds.yaml specified in the cluster deployment.
The openstack actuator now reads the CertificatesSecretRef to generate a concatenated trust bundle
and update the corresponding clouds.yaml entry with the trust.

The gopercloud/utils lib supports setting the CACertFile with path to the CA or the CA contents itself,
and current we set the CA directly as bytes to reduce the write to file -> set the file path -> cleanup the
temp file.