hivev1 validating-webhooks: allow adding checks for fields in clusterdeployment based on feature gates
the webhook looks for feature gates enabled as a comma separate list of feature gates
set as env HIVE_FEATURE_GATES_ENABLED

Adds 2 methods to help validating
- making sure certain field is set only when a feature gate was enabled
- maling sure certain field is might be equal in value to certain value only when
  a feature gate was enabled.
These help the most common use case of making sure certain platforms are only set when
the alpah feature gates are set.

The webhook reads the env for every request so that it can respond to changing feature gates
without reintialization.
If validation fails due to missing feature gate, the object is immediately rejected to prevent
running any code later on that assumes features gates to be set.