hive operator: sync featuregate config map and hive admission to use it
the operator now creates and manages a config map hive-feature-gate to allow
operands to fetch which feature gates are enabled. The operands can load the
configmap as env using
```yaml
envFrom:
  configMapRef:
    name: hive-feature-gate
```
Currently the k8s api doesn't support auto updating the env when the config map
changes, see [1], so the operator stores the hash of the config map data on the
deployment as annotation. This ensures that the deployment rolls when the data
changes in the config map.

Currently only the hive-admission deployment is wired to get the feature gates.

[1]: https://kubernetes.io/docs/concepts/configuration/configmap/#mounted-configmaps-are-updated-automatically