Point sigs.k8s.io/controller-runtime to a fork
In an effort to reduce the storage for nonessential memory-intensive
metrics, rest_client_request_latency_seconds was identified as the
metric that reports and stores the most time series. This metric was
vendored in from sigs.k8s.io/controller-runtime.
However, this metric has been deprecated since kube 1.14 and turned off
in kube 1.17. Since upstream still registers this metric, it was
consequently decided to point controller-runtime to a fork that carries
the patch to not report this metric