installpods: allow using AWS config in install/uninstall
Update the AWS provision and deprovions pods to

- mount the credentials secret to /etc/aws

- set env variables for loading aws cli shared config from
  /etc/aws/aws_config

This means that when the credentials secret has aws cli config in
aws_config key, it will be used by the sdk for creating the clients.

This removes the restriction of setting aws_access_key_id and
aws_access_secret_key in the secret.