awstagdeprovision: support AWS config while keeping refresh function
Previously the awstagdeprovision would accept static creds from Env or
from fixed file locations. This means even when AWS CLI Config was being
provided using env, it would not accept that.

So now, it adds a cred-auth-dir flag like the other deprovisioners and
watches for changes in that dir. This free up the restriction
credentials source and allows using all the sources like the provision
pod.