Fix clusters stuck resuming due to stopped instance.
x-ref: https://issues.redhat.com/browse/HIVE-1455

Encountered a cluster deployment stuck resuming for several days, and
thus was not powered off after the hibernateAfter interval.

Issue appeared to be because one worker VM was still in stopped state.
The code made the assumption that we must have issued the start command
when transitioning to resuming state, and then never tries to again.

To fix we now will issue StartMachines each time we see that not all
machine are running yet.
All three providers are implemented to only issue to VMs that
are stopped / stopping so the change should have no impact normally.