Bug 1947220: Skip uninstalled clusters in clustersync controller.
This was happening previously for most clusters due to an assumption
that if no Unreachable condition exists, then the cluster must not be
installed yet. However an integrating application stumbled into a crash
where they were misusing the Unreachable condition, with status Unknown,
and our code interpreted this as must be installed (it wasn't), and
reachable, leading to the crash as no cluster metadata is defined.

The misused of Unknown condition will be fixed on their end, but we
should have an explicit installed check regardless.