hive-operator: resync on changes to secrets in HiveConfig
When certain secrets referenced in HiveConfig change, the operator
should retriger the reconcile.

Currently only secrets in `spec.additionalCertificateAuthoritiesSecretRef`
support this behavior.

hive operator maintains an index of secrets referenced in HiveConfig
that can trigger reconcile. When the operator sees an update for these
secrets, it uses the index to trigger reconcile for the specific
HiveConfig.