add clusterinstall contract
Adds a Go type for the interacting with data from implementations of the
clusterinstall contract. This type allows controllers to fetch the
implementation in json form and then marshal that into the Go type. This
enables the controllers to interact with a defined-type instead of
unstructured.

The Go type also defines the clusterinstall contract's json marshalled
form.

Adds a clusterInstallRef to ClusterDeployment, that allows users to
point to an implementation of the clusterinstall contract using
Group,Version,Kind and name. This cannot be set with .spec.provisioning
field on ClusterDeployment.