operator: list CRDs in cluster to compute supported hive contracts
Hive operator now watches for changes to CRDs. And when CRDs change it
reconciles all the HiveConfig objects to calculate the new list of
supported contracts

Hive operator uses the list of CRDs in the cluster to generate a list of
supported hive contracts and all their implementations.

If the CRD has label `contracts.hive.openshift.io/<contract name>`, it
adds all the versions from the crd as supported implementation of the
contract.

The operator creates a configmap hive-support-contracts with this list
of supported implementations for various contracts. To prevent users
from adding contracts to CRDs that hive then accepts, the operator keeps
a list of allowed contracts.

The configmap is mounted in the hive-admission pods so that it can use
that list to decide whether the referenced objects are infact allowed.