operator: use the correct cache to fetch secrets in hive ns
When we receive a watch, and we reconcile for that watch and then use
that cache to get the latest version of the changed object, we can
assume this gurantee.

Currently, due to historical reasons, we have a separate cache for
watching secrets in hive namespace, but then use the reconciler's client
to get the secrets. This means that our assumption of having the latest
update secret is no longer valid.

So in the current setup, we should use the lister from the correct
cache to get secrets. This is now achieved by storing the lister for the
secrets and then using that when required.