Hide massive PodSpec definition in ClusterProvision CRD yaml.
This PodSpec alone is taking up 350K of the hard 1024 limit currently
imposed on OLM bundles.

We do not consider PodSpec an integral part of this API and more an
implementation detail, it is extremely unlikely anyone is relying on
being documented in the CRD.

x-ref: ttps://issues.redhat.com/browse/HIVE-1513

Because we still use v1beta1 CRDs, preserveUnknownFields is default.
When we switch to v1 we will need preserveUnknown on the
ClusterProvisionSpec to keep the PodSpec maintained. This is noted in
the relevant Jira.