remotemachinesets: sync ready replicas and errors to machinepool
Add ReadyReplicas and Error{Reason, Message} to status for MachineSets
in the MachinePool. This allows the machinepool to provider more
accurate status of important information of the machinesets in the remote
cluster without having to probe into the cluster itself.

There is still a caveat the the machinepool doesn't react to the
machinesets in the remote cluster, so this status will not be latest
unless the machinepool changed or other updates caused a re sync.