Fix stuck privatelink clusters deleted prior to install.
In some scenarios when an AWS privatelink cluster is deleted prior to
being fully installed, it was possible for the code to jam up in a
situation where no hosted zone ID was recorded in status, and we cannot
discover it because the admin kubeconfig secret is already gone.

To fix we just gracefully exit in this scenario as we have nothing to
cleanup.