Sort clusterdeployment conditions
Now that we have all possible cluster deployment conditions present on
the CR, it can become very difficult to debug.
This commit tries to sort the list of conditions, such that the
conditions in their undesired state are always at the top of the list,
followed by conditions in their desired state. Conditions with status
unknown are present at the bottom of the list.
Each list of undesired, desired and unknown status conditions are sorted
by their type