Validate region in InstallConfig matches ClusterDeployment.
In general we try not to fully validate install configs match what you
set in ClusterDeployment. In this particular case however, a region
mismatch will cause the cluster to keep provisioning machines in the
wrong region until your cloud quota is fully exhausted.

Add a minimal check to ensure the InstallConfig region matches the
cluster deployment for AWS, Azure, and GCP.

In the event of a problem, a RequirementsMet condition on the
ClusterDeployment will go False. Hoping to use this condition to
consolidate a few others in this controller that are failing pre-flight
checks.